+++ common/src/main/java/net/conczin/immersive_furniture/config/Config.java
@@ -40,6 +40,7 @@ public final class Config extends JsonConfig {
 
     // Atlas size for baked furniture; must be a multiple of 2, between 512 and 8192
     public int bakedAtlasSize = 1024;
+    public boolean disableAtlasRefreshes = false;
 
     public Config(String name) {
         super(name);
@@ -51,8 +52,13 @@ public final class Config extends JsonConfig {
 
     public int getBakedAtlasSize() {
         return switch (bakedAtlasSize) {
-            case 512, 1024, 2048, 4096, 8192 -> bakedAtlasSize;
+            case 256, 512, 1024, 2048, 4096, 8192 -> bakedAtlasSize;
             default -> 1024;
         };
+    }
+
+    @Override
+    int getVersion() {
+        return 2;
     }
 }